pyproject.toml
[tool.poetry]
name = "coby-test-book"
version = "0.1.0"
description = ""
authors = ["coby@vessl.ai"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymupdf = "^1.25.5"
pandas = "^2.2.3"
openai = "^1.70.0"
python-dotenv = "^1.1.0"
tqdm = "^4.67.1"
scikit-learn = "^1.6.1"
transformers = "^4.51.0"
torch = "^2.2.0"
accelerate = "^0.27.2"
bitsandbytes = "^0.42.0"
peft = "^0.9.0"
poetry-plugin-export = "^1.9.0"
datasets = "^2.18.0"
scipy = "^1.12.0"
sentencepiece = "^0.2.0"
nltk = "^3.8.1"
rouge-score = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
